44 Chain Syntax Grammar
Added the support for text in the EBNF and added all the rules (including the keywords) within the main EBNF
Removing the need to support special keywords
Removing the need and optimization to parse keywords first
We can know parse the whole input as through the grammar.

Next todo:
- Add the mergeable keywords in the ebnf rule
- Detect the rules with no parents (in the debug UI and comment them)
- Detect the rules with no children that are not leaves (in the debug UI and comment them)
- Then redo the compiler (again) to only get the rules from the ebnf grammar AND the custom dll, if any (for now theory only).
- Then start writing the parser and the error handling from the grammar
