+++ Grammar UI/EbnfParserViewModel.cs
@@ -2,10 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Ebnf;
 
 namespace Grammar_UI
